ml-service/requirements.txt
Flask==3.0.0
tensorflow
scikit-learn
pandas
numpy
Pillow
huggingface_hub
kaggle
python-dotenv
tf-keras
